Replace useless tests with useful ones
